requirements.txt
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
peft>=0.7.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
gradio>=4.0.0
tqdm>=4.65.0
jieba>=0.42.1
datasets>=2.14.0
tensorboard>=2.14.0
ultralytics>=8.0.0
clip@git+https://github.com/openai/CLIP.git